--- Sources/client/Constants.swift
@@ -1,65 +1,0 @@
-// Copyright (c) 2019 David Stump <david@davidstump.net>
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-
-/// Default timeout when making a connection set to 10 seconds
-public let PHOENIX_DEFAULT_TIMEOUT: Int = 10000
-
-// Default timeout when making a connetion, set to 10.0 seconds
-public let PHOENIX_TIMEOUT_INTERVAL: TimeInterval = 10.0
-
-/// Default heartbeat interval set to 30 seconds
-public let PHOENIX_DEFAULT_HEARTBEAT: Int = 30000
-
-/// Default heartbeat interval set to 30.0 seconds
-public let PHOENIX_HEARTBEAT_INTERVAL: TimeInterval = 30.0
-
-
-
-/// Represents the multiple states that a Channel can be in
-/// throughout it's lifecycle.
-public enum ChannelState: String {
-    case closed = "closed"
-    case errored = "errored"
-    case joined = "joined"
-    case joining = "joining"
-    case leaving = "leaving"
-}
-
-/// Represents the different events that can be sent through
-/// a channel regarding a Channel's lifecycle.
-public struct ChannelEvent {
-    public static let heartbeat = "heartbeat"
-    public static let join      = "phx_join"
-    public static let leave     = "phx_leave"
-    public static let reply     = "phx_reply"
-    public static let error     = "phx_error"
-    public static let close     = "phx_close"
-    
-    static func isLifecyleEvent(_ event: String) -> Bool {
-        switch event {
-        case join, leave, reply, error, close: return true
-        default: return false
-        }
-    }
-}
-
-
-
