+++ Sources/client/Supporting Files/Info.plist
