+++ Sources/client/Supporting Files/SwiftPhoenixClient.h
